pyproject.toml
[project]
name = "fastapi-redis-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.37",
]
